// src/app/models/attention-card.model.ts
export class AttentionCardModel{
    id: any;
    clinichistory: number;
    dateadmission?: string;
    motivo: string;
    ma: boolean;
    mmp: boolean;
    mp: boolean;
    ar: boolean;
    dcb: boolean;
    dtm: boolean;
    mm: boolean;
    af: boolean;
    asc: boolean;
    state?: number;
    createdAt?: Date;
    updatedAt?: Date;
}